refactor: norm対応。可読性の改善。
#50
